Introduce validatePostalCode function and catch unsupported locales
