+++ web/app/containers/account-address/partials/address-form.jsx
@@ -6,13 +6,14 @@ import React from 'react'
 import {connect} from 'react-redux'
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import * as ReduxForm from 'redux-form'
-import isPostalCode from 'validator/lib/isPostalCode'
+
+import Button from 'progressive-web-sdk/dist/components/button'
+import {UI_NAME} from 'progressive-web-sdk/dist/analytics/data-objects/'
 
 import {ADDRESS_FORM_NAME} from '../../../store/form/constants'
-import Button from 'progressive-web-sdk/dist/components/button'
+import {validatePostalCode} from '../../../utils/validation'
 import {getIsEditing} from '../selectors'
 import {submitAddAddress, submitEditAddress} from '../actions'
-import {UI_NAME} from 'progressive-web-sdk/dist/analytics/data-objects/'
 import AccountAddressFields from './account-address-fields'
 
 const REQUIRED_TEXT = 'Required'
@@ -30,7 +31,7 @@ const validate = (values, props) => {
         'telephone'
     ]
 
-    if (values.countryId && values.postcode && !isPostalCode(values.postcode, values.countryId.toUpperCase())) {
+    if (values.countryId && values.postcode && !validatePostalCode(values.postcode, values.countryId.toUpperCase())) {
         errors.postcode = 'Enter a valid postal code'
     }
