+++ web/app/containers/checkout-shipping/partials/checkout-shipping-form.jsx
@@ -7,9 +7,8 @@ import {connect} from 'react-redux'
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import * as ReduxForm from 'redux-form'
 import isEmail from 'validator/lib/isEmail'
-import isPostalCode from 'validator/lib/isPostalCode'
 
-import {validateFullName} from '../../../utils/utils'
+import {validateFullName, validatePostalCode} from '../../../utils/validation'
 import {getIsLoggedIn} from 'progressive-web-sdk/dist/store/user/selectors'
 import {getInitialShippingAddress} from '../../../store/checkout/shipping/selectors'
 
@@ -54,7 +53,7 @@ const validate = (values, props) => {
         errors.username = REQUIRED_TEXT
     }
 
-    if (values.countryId && values.postcode && !isPostalCode(values.postcode, values.countryId.toUpperCase())) {
+    if (values.countryId && values.postcode && !validatePostalCode(values.postcode, values.countryId.toUpperCase())) {
         errors.postcode = 'Enter a valid postal code' // or zip code?
     }
